Updated polarity to now divide by the square root of n and to stop stripping the apostrophes from the text.
The issue of not really affordable vs really not affordable and weighting them differently appears to be logical at first.  However, qdap is meant to analyze dsicourse.  Depending on how words are stressed depends on their meaning.  For that reason it is not possible to make an algorithm that discriminates between the stressed word use based on the transcripts along.  I will maintain the algorith assigning equaly polarity scores to both of these sentences.

close #136